+++ src/policies/no_sampling_policy.cc
@@ -4,7 +4,7 @@ NoSamplingPolicy::NoSamplingPolicy() {}
 
 NoSamplingPolicy::~NoSamplingPolicy() {}
 
-bool NoSamplingPolicy::accept(std::uint64_t timestamp, float* value)
+bool NoSamplingPolicy::accept(std::uint64_t timestamp, double* value)
 {
     return true;
 }
